% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabias-package.R
\docType{package}
\name{metabias-package}
\alias{metabias-package}
\alias{_PACKAGE}
\title{metabias: Meta-Analysis for Within-Study and/or Across-Study Biases}
\description{
Provides common components (classes, methods, documentation) for packages that conduct meta-analytic corrections and sensitivity analyses for within-study and/or across-study biases in meta-analysis. See the packages `PublicationBias`, `phacking`, and `multibiasmeta`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikabr/metabias}
  \item Report bugs at \url{https://github.com/mikabr/metabias/issues}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Maya Mathur \email{mmathur@stanford.edu}
}

}
\keyword{internal}
