% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabias-class.R
\name{new_metabias}
\alias{new_metabias}
\title{Constructor for metabias S3 class}
\usage{
new_metabias(x = list())
}
\arguments{
\item{x}{List with elements "data", "values", "stats", "fits"}
}
\description{
Constructor for metabias S3 class
}
