% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabias-class.R, R/methods.R
\name{metabias-class}
\alias{metabias-class}
\alias{metabias}
\alias{print.metabias}
\alias{summary.metabias}
\title{metabias S3 class}
\usage{
metabias(
  data = data.frame(),
  values = list(),
  stats = data.frame(),
  fits = list()
)

\method{print}{metabias}(x, ...)

\method{summary}{metabias}(object, ...)
}
\arguments{
\item{data}{Dataframe containing data used to fit the model(s), with added
columns for any values computed during model fitting.}

\item{values}{List of values of arguments passed to the function.}

\item{stats}{Dataframe of summary statistics from the model fit(s).}

\item{fits}{List of fitted objects (which have a class that depends on the
underlying fitting methods, e.g. \code{\link[robumeta:robu]{robumeta::robu}} or \code{\link[rstan:stanfit-class]{rstan::stanfit}}).}

\item{x}{Object of class \code{metabias}.}

\item{...}{Not used.}

\item{object}{Object of class \code{metabias}.}
}
\description{
A object of class \code{metabias} is the result of fitting one or
more models to a dataset with one row per study being meta-analyzed. These
models are either (1) a meta-analysis with a correction for one or more
within-study or across-study biases, or (2) a sensitivity analysis for
meta-analyses with respect to these biases. Examples of functions that
return such objects include:
\itemize{
\item \code{PublicationBias::pubbias_meta()}
\item \code{PublicationBias::pubbias_svalue()}
\item \code{phacking::phacking_meta()}
\item \code{multibiasmeta::multibias_meta()}
\item \code{multibiasmeta::multibias_evalue()}
}
}
